The sum of two integers in a loop
A program that reads two numbers in a loop and prints the sum in the standard output. If a user enters only a single number, the program should print the same number. If a user enters an empty line, the program should ignore it.
When the command "/exit" is entered, the program must print "Bye!" and then stop.